more cycle refactoring
