+++ packages/cursorless-engine/src/processTargets/targets/UntypedTarget.ts
@@ -1,12 +1,12 @@
 import { Range } from "@cursorless/common";
 import { BaseTarget, CommonTargetParameters } from "./BaseTarget";
 import type { Target } from "../../typings/target.types";
-import {createContinuousRangeUntypedTarget} from "../targetUtil/createContinuousRangeUntypedTarget";
 import {
   getTokenLeadingDelimiterTarget,
   getTokenRemovalRange,
   getTokenTrailingDelimiterTarget,
 } from "../targetUtil/insertionRemovalBehaviors/TokenInsertionRemovalBehavior";
+import {createContinuousRange} from "../targetUtil/createContinuousRange";
 
 interface UntypedTargetParameters extends CommonTargetParameters {
   readonly hasExplicitRange: boolean;
@@ -65,3 +65,24 @@ export class UntypedTarget extends BaseTarget<UntypedTargetParameters> {
     };
   }
 }
+export function createContinuousRangeUntypedTarget(
+  isReversed: boolean,
+  startTarget: Target,
+  endTarget: Target,
+  includeStart: boolean,
+  includeEnd: boolean
+): UntypedTarget {
+  return new UntypedTarget({
+    editor: startTarget.editor,
+    isReversed,
+    hasExplicitRange: true,
+    contentRange: createContinuousRange(
+      startTarget,
+      endTarget,
+      includeStart,
+      includeEnd
+    ),
+    isToken: startTarget.isToken && endTarget.isToken,
+  });
+}
+
